Refactored name for better semantic clarity.
