+++ lmcache/v1/multiprocess/adapter_connector/base.py
