+++ lmcache/v1/multiprocess/adapter_connector/pickle.py
@@ -9,11 +9,11 @@ import pickle
 import torch
 
 # First Party
-from lmcache.v1.multiprocess.protocol import RequestType, get_response_class
-from lmcache.v1.multiprocess.worker_transfer.base import (
+from lmcache.v1.multiprocess.adapter_connector.base import (
     NonGpuContext,
     NonGpuContextMetadata,
 )
+from lmcache.v1.multiprocess.protocol import RequestType, get_response_class
 
 
 class NonGpuContextPickle(NonGpuContext):
