+++ lmcache/v1/multiprocess/adapter_connector/shm.py
@@ -10,12 +10,12 @@ from typing import Any
 import torch
 
 # First Party
-from lmcache.v1.multiprocess.protocol import RequestType, get_response_class
-from lmcache.v1.multiprocess.worker_transfer.base import (
+from lmcache.v1.multiprocess.adapter_connector.base import (
     NonGpuContext,
     NonGpuContextMetadata,
 )
-from lmcache.v1.multiprocess.worker_transfer.shm_types import ShmSlotDescriptor
+from lmcache.v1.multiprocess.adapter_connector.shm_types import ShmSlotDescriptor
+from lmcache.v1.multiprocess.protocol import RequestType, get_response_class
 
 
 class NonGpuContextShm(NonGpuContext):
