+++ lmcache/v1/multiprocess/adapter_connector/worker_transfer.py
@@ -13,12 +13,7 @@ from lmcache import torch_dev
 from lmcache.utils import EngineType, init_logger
 from lmcache.v1.distributed.api import MemoryLayoutDesc
 from lmcache.v1.gpu_connector.utils import LayoutHints, is_mla
-from lmcache.v1.multiprocess.custom_types import RegisterNonGpuContextPayload
-from lmcache.v1.multiprocess.futures import MessagingFuture
-from lmcache.v1.multiprocess.mq import MessageQueueClient
-from lmcache.v1.multiprocess.protocol import RequestType
-from lmcache.v1.multiprocess.protocols.engine import RegisterNonGpuContextResponse
-from lmcache.v1.multiprocess.worker_transfer.base import (
+from lmcache.v1.multiprocess.adapter_connector.base import (
     NonGpuContext,
     NonGpuContextMetadata,
     compute_kv_layout,
@@ -26,6 +21,11 @@ from lmcache.v1.multiprocess.worker_transfer.base import (
     gather_paged_kv_to_cpu,
     scatter_cpu_to_paged_kv,
 )
+from lmcache.v1.multiprocess.custom_types import RegisterNonGpuContextPayload
+from lmcache.v1.multiprocess.futures import MessagingFuture
+from lmcache.v1.multiprocess.mq import MessageQueueClient
+from lmcache.v1.multiprocess.protocol import RequestType
+from lmcache.v1.multiprocess.protocols.engine import RegisterNonGpuContextResponse
 
 logger = init_logger(__name__)
