+++ lmcache/v1/multiprocess/modules/non_gpu_transfer.py
@@ -15,6 +15,7 @@ from lmcache.v1.distributed.api import (
     MemoryLayoutDesc,
     ObjectKey,
 )
+from lmcache.v1.multiprocess.adapter_connector.base import NonGpuContextMetadata
 from lmcache.v1.multiprocess.custom_types import (
     IPCCacheEngineKey,
     RegisterNonGpuContextPayload,
@@ -30,7 +31,6 @@ from lmcache.v1.multiprocess.protocols.engine import (
     PrepareStoreResponse,
     RegisterNonGpuContextResponse,
 )
-from lmcache.v1.multiprocess.worker_transfer.base import NonGpuContextMetadata
 
 # Local
 from .server_transfer import (
